Refactor: 로그인 페이지 SOLID 원칙에 맞게 리팩토링
